+++ CHANGELOG.md
@@ -8,6 +8,7 @@ ADDED:
 
 - new decorator class for JMSpinner that encapsulate a property model and bind it to it
 - new methods in class RobotExtensions for mouse move with argument of thread priority
+- new method in ClipboardExtensions for copy a given string to the clipboard
 
 Version 5.9
 -------------